requirements.txt
aiohttp==3.9.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==3.7.1
async-timeout==4.0.3
attrs==25.3.0
bcrypt==4.3.0
beautifulsoup4==4.12.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
coverage==7.10.2
cryptography==45.0.6
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.103.1
frozenlist==1.7.0
grpcio==1.74.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==0.17.3
httpx==0.24.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
loguru==0.7.3
multidict==6.6.3
numpy==2.0.2
openai==1.3.0
packaging==25.0
pandas==2.3.1
passlib==1.7.4
pinecone-client==2.2.4
pluggy==1.6.0
portalocker==3.2.0
propcache==0.3.2
protobuf==6.31.1
psycopg2-binary==2.9.7
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.1
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.3.0
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.2
qdrant-client==1.15.1
requests==2.31.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.21
starlette==0.27.0
tomli==2.2.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.23.2
yarl==1.20.1
